requirements.txt
matplotlib==1.4.3
mock==1.0.1
nose==1.3.6
numpy==1.9.2
pydot2==1.0.33
pyparsing==1.5.7
python-dateutil==2.4.2
pytz==2015.2
scikit-learn==0.16.1
scipy==0.15.1
six==1.9.0
